requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
python-multipart
PyPDF2
python-docx
Pillow
pytesseract
groq
